Library (fits, copykeys): List of Keywords
Until now, copykeys option for astfits used to only accept a range of
keywords for eg: 5:10.

With this commit it now also accepts a range of keywords as a comma
separated list for eg: BITPIX,LTV1.

This is done by parsing the input of copykeys, and if an alphabet is
detected, then it considers them as a list of keywords.

This is a solution to the task#16073.