radial-profile: new option for considering zero-valued pixels
Until now, in the radial profile script the zero-valued pixels were always
considered as blank. Because in cropping step these pixels were set to
blank. Therefore, there was not any way for considering these pixels in the
radial profile.

With this commit, a new option has been added to the radial profile script
for this purpose: '--zeroisnotblank'. If this option is used, the same
option is added to the crop command.

This task was proposed by Ignacio Ruiz Cejudo.