Development: only Makefile.in files in .gitignore
Until now, '.gitignore' was set to generally ignore '*.in' (all files
ending in '.in'). The main reason was the 'Makefile.in' files that are
created during the bootstrapping. However, this caused a problem of
forgetting to add the raw installed script files: they weren't
automatically shown in 'git status', or we needed to do a '--force' with
'git add'.

With this commit, only 'Makefile.in' files are ignored and all other '*.in'
files will be shown by git. While checking a clean build of Gnuastro in the
source directory, I noticed that the newly added 'tests/programs-built'
directory (built during 'make check') also needs to be ignored, so its also
added.

I also noticed that the 'bin/scripts/fits-view.in' file had an executable
flag, so its executable flag is now removed.

This issue was reported by Sepideh Eskandarlou.