MakeProfiles: new option to insert custom image
Until now, MakeProfiles created the pixels of each profile internally and
placed them in the image. However, the internal way it done this was to
build an image for each profile separately (in parallel!) and put them into
the image. Therefore a natural extension was to let the user provide a
custom image to place in the desired coordinates.

With this commit, it is now possible to do just that with a new type of
profile ('custom-img') and two options: '--customimg' and
'--customimghdu'. The user can give any number of custom images to these
options and call them by their ID in the "radius" column.