Book: Example for Magnitude Limit
Until now, there wasn't any example or guide for calculating the magnitude
in the section of "Magnitude limit of image".

With this commit, we bring up an example in this section using required
commands that can give a detailed oriented picture on this section. For
this aim, we also take advantage of "astscript-ds9-region" to show the
result on the images.

Thanks to Sepideh Eskandarlou and Mohammad Akhlaghi for helping out to
underestand and complete this section.