Library (checkset): also using MemFree to find available memory
Until now, to find the available memory, Gnuastro's memory allocation
library would only check the 'MemAvailable' keyword in
'/proc/meminfo'. However, on some systems (tested on Ubuntu within the
Windows Subsystem for Linux), only has the 'MemFree' keyword is
available. On most other native GNU/Linux systems, both these keywords are
available.

Therefore, when reading a large file on the former type of operating
systems, Gnuastro programs will print a warning (and would not be able to
find how much available memory there is!). When there was no problem (space
was actually available), the program would finish, but this warning would
be annoying for the user. In a worst case, this inability would lead to a
crash (due to a lack of memory).

With this commit, the part that reads the keyword and its value is now
separated (modularized/abstracted) into a separate function. Thanks to the
newly spinned-off function, Gnuastro now checks for both keywords.

In fact, after reading more about the difference between these keywords (in
[1]), I noticed that 'MemFree' is the actual physical available memory,
while 'MemAvaialble' is the estimated (not necessarily actually free)
space. So 'MemFree' is actually the better keyword to check first. As a
result, we now check 'MemFree' and if it doesn't exist, we'll check
'MemAvailable'.

On a sidenote, I noticed that the spacing between lines in one part of
'bin/mkcatalog/ui.c' was not correct, so I added a new empty line (this was
not worth a separate commit, so I am committing it with this!).

This bug was reported by Juan Miro.

This fixes bug #62305.

[1] https://lynxbee.com/understanding-proc-meminfo-analyzing-linux-memory-utilization/#.YlYJgR5ByXI