Book: unified method of calculating area in arcmin2
Until now, throughout the "coverage" parts of the second tutorial ("Angular
coverage on the sky" and "Cosmological coverage" sections), we had used the
format of '(X*60)^2' to calculate the area from arcsec-united edges to
arcmin^2. However, in the last example of the "Cosmological coverage"
section, we had used '(X^2 * 3600)'. While this is mathematically
identical, the different ways of doing the calculation could be confusing.

With this commit, the last usage has been edited to be similar to the rest
and help in the clarity of the text.

This was reported by Sílvia Farras.