MakeCatalog: added check to avoid objects images with negative values
Until now, when MakeCatalog was given an input (labeled) image that
contained negative values, it would crash with a segmentation fault
(because the negative pixels in an un-signed array index would lead to
extremely large values).

With this commit, on the first time that we parse the input object labels,
if we find a non-blank negative pixels in the image, an informative warning
is printed and MakeCatalog aborts cleanly.

This bug was reported by Sepideh Eskandarlou.

This fixes bug #62309.