MakeProfiles: better column metadata for 3D test
Until now, the only column with metadata in the 'tests/mkprof/3d-cat.txt'
file (that is fed into MakeProfiles to build the input 3D cube) was the
fifth column (to enable having strings as function identifiers). This was
not convenient/clear.

With this commit, all the columns are given metadata.