Webpage: added link to ADASS tutorial, old citations removed
Until now, the video of the Gnuastro tutorial in ADASS 2021 wasn't linked
on the top Gnuastro webpage. Also, the list of citations to Gnuastro had
become old (last updated in September 2020!).

With this commit, the ADASS video link has been added and the
hard-to-update list of citations has been removed, keeping the links to ADS
and Google Scholar to guide interested readers.