Checks: dependencies of two astscript-psf scripts corrected
Until now, the dependencies of the 'make check' tests of all the PSF
scripts was 'mkprof/mosaic2.sh.log'. However, for two of the scripts, the
input image that was used was actually the output of 'mkprof/mosaic1.sh'!
This caused random crashes some times (when 'mosaic1.sh' wasn't yet run).

With this commit, these two particular scripts now depend on the
'mkprof/mosaic1.sh.log'.