Book: recommending Codeberg instead of Gitlab for forks
Until now, in multiple places of the book, we were recommending users to
use Gitlab (instead of Github, which is the most popular, but non-free
software). At the time of writing, Gitlab was the best alternative for
forks (not including Savannah which seems to be primarily for the main
project repository, not forks). However, in the last few years, I have
learnt about Codeberg which is fully free software (as opposed to Gitlab
which has a proprietary component). Furthermore, Gitlab is not accessible
in many countries, making it hard to collaborate.

With this commit, we now recommend to use Codeberg in the relevant parts of
the book.

In the process, I also noticed that those older parts of the book hadn't
yet been converted to a one line per sentence format like the rest of the
book (which greatly helps in tracking updates in long texts with Git). So
their formatting is also corrected with this commit.