Book: new section in second tutorial on measuring limits
Until now, the second tutorial didn't have any examples on how to measure
some very important properties of the dataset: the limits (detection,
completeness or surface brightness).

Following the nice examples of the last few commits by Zahra (and edited by
Sepideh), we thought it would be best to blend that discussion into the
second tutorial where we datasets are ready for the reader to actually be
able to run the commands and help in making the arguments.