.mailmap: corrected to use single name for Zahra Hosseini
In an earlier commit, Zahra's commit author was named as "Zahra Hosseini
Shahisavandi", however, in a more recent commit, the name "S. Zahra
Hosseini Shahisavandi" was used. This would result in an in-correct author
count when generating the 'AUTHORS' file or the second page list of authors
in the PDF book.

With this commit, to avoid this, Zahra's email has been registered in Git's
'.mailmap' file, so it recognizes both with the same name (in the latter
format).