Book: new item on finding counts at different exposure times
Until now, in the "Brightness, Flux, Magnitude and Surface brightness"
section, we were always assuming that the data are in units of
counts/sec. However, when trying to simulate images of raw exposures at
differing exposure times, we noticed that there was no discussion on how to
account for varying exposure time.

With this commit, a new item has been added under the zeropoint, describing
how to account for the exposure time to produce an image based on counts,
or how to derive an exposure-time-dependent zeropoint to directly use in
MakeProfiles or MakeCatlog.

This was done with the help of Raul Infante-Sainz and Sepideh Eskandarlou.