Library (checkset): only using MemFree when MemAvailable not present
Until now, when we wanted to find the amount of available RAM, we would
first search for 'MemFree', and then 'MemAvailable'. This was because
previously we had noticed that some systems (Ubuntu, on Windows Subsystem
for Linux) don't have 'MemAvailable'!

However, when both are present, 'MemFree' is usually much smaller than
'MemAvailable'. As a result, Gnuastro would use memory mapping, even when
there was available memory!

With this commit, we now read both, but only use 'MemFree', when
'MemAvailable' couldn't be read.

This issue was reported by Raul Infante-Sainz.