MakeProfiles: now using input catalog's magnitude column by default
Until now, when a custom profile was given to MakeProfiles, the profile's
raw values were directly inserted into the 2D image (effectively ignoring
the requested magnitude in the main input catalog!). This was un-expected
behavior (because the magnitude should be set in that high-level catalog;
and this works for all other profile types).

With this commit, by default once the 2D image is created, it will be
scaled to the requested magnitude in that particular row of the input
catalog (like all other profiles!). However, if a user wants to use the raw
values in the profile (and ignore the total magnitude), they can use the
newly added '--mcolnocustprof' option to get the old behavior.

This was suggested by Elham Saremi.