Segment: --quiet option is now fully quiet (was missing a printed line)
Until now, when Segment was called with the '--quiet' option, it would
mistakenly print the "- Finding true clumps..." output line (while every
other output line was supressed). This was happening because I had forgot
to add a check for quiet-mode before printing this!

With this commit, the condition has been added, so Segment is now fully
quiet when the user gives '--quiet'.

This bug was reported by Teet Kuutma.

This fixes bug #62429.