Book: several corrections in Sufi tutorial and MakeProfiles
Until now, in the Sufi tutorial the name of the output kernel was
incorrectly used as '0_cat.fits'. This was the default name for older
versions of MakeProfiles. The current MakeProfiles produces a file called
'0_cat_profiles.fits'. There were also several typos in the MakeProfiles
section of the book.

In this commit, both issues have been corrected.