Book: corrected typo in script at the end of the Sufi tutorial
Until now, in the script of the end of the Sufi tutorial, we were using the
old convention for the mock kernel that was created
('0_"$base".fits'). However, as described in the previous commit, newer
versions of MakeProfiles use a '_profiles.fits' suffix, not just '.fits').

With this commit, this issue has been corrected.

This issue was reported by Faezeh Bijarchian.