Book: full edit of first tutorial, with redirection to make catalog
Until now, in the first tutorial, we were simply asking the user to
copy-paste the contents of the table (to produce the necessary mock image)
using a plain-text editor. However, the series of commands were
confusing. Also, there was only metadata for one column (which could again
confuse a reader!).

With this commit, we now describe metadata, add metadata for all the
columns and use shell re-direction to automatically create the table. In
the process, I also reviewed the whole tutorial and added some more modern
features like using 'astscript-fits-view'.

Also, until now, the "Getting involved" part of the main webpage had a
sub-section on the Google Summer of Code (GSoC) 2022. But GSoC 2022
application period has already finished, so there is no more need to have
that entry in the webpage for this year.

With this commit, that part of the webpage has been commented, so we can
use it again for next year if necessary.

The issues in the first tutorial were reported by Marjan Akbari.