Library (jpeg.h and tiff.h): addressed compiler warning for const
Until now, when the host operating system didn't have the JPEG or TIFF
libraries, during the 'make' phase, we would see compiler warnings saying
that a 'const' is missing in the 'jpeg_error_no_libjpeg' and
'tiff_error_no_litiff' functions. This happened because we were passing the
caller function's name to them (through '__func__'). However, '__func__' is
a constant string defined at compile time, so to be clean (and avoid
warnings), its better to put a 'const' before its definition.

With this commit, a 'const' qualifier has been placed before the argument
definition in these two functions and this removed the compiler warning.

This issue was found with the help of Jeremy Lim.