Library (txt.h): accounts for larger width in str col (when its last)
Until now, when the last column of an input table (in plain-text) has a
string type, but the given width (number of characters) in the column's
metadata gives a width that is larger than the string, Table will not
recognize that string column! Therefore, it will either not read that
column if there are other columns before it, or if that string column is
the only column, it will complain that "no usable data rows".

With this commit, the solution was found to be a wrongly placed "break" in
cases where the expected width of a string column (that is also the last
column!) is larger than the actual characters of the last string column
(before reaching the new-line character). The check to 'break' the loop
should have been moved to the start of the 'while' loop, not in the same
step that the line is being parsed.

This bug was found with the help of Sepideh Eskandarlou.

This fixes bug #62548.