Configuration: automatically obtaining the number of threads
Until now, when informing the user the number of cores to be used in the
make step ('make' and 'make check'), we simply recommended using
'-j8'. However, depending on the particular machine, the number of threads
could be higher or lower.

With this commit, the number of available threads is automatically computed
in the './configure' phase. The way of obtaining it is different depending
on the system (GNU/Linux or macOS). Once it has been computed, in the
messages, that value is used after '-j' to help guide the user.