make check: using awk instead of find to look for script filenames
Until now, the 'find' program was used to obtain the name of the
(executable) installed scripts (through the bit-flags of a file). After
that, a symbolic link was created to have them in the desired directory for
testing. However, this way of obtaining the scripts does not work on macOS
systems (tested on Monterey version 12.3).

With this commit, the way of obtaining these filenames has changed. Now
they are obtained using 'awk' and by parsing their names directly from the
'bin_SCRIPTS' variable of 'Makefile.am' file. Later, a symbolic link to
each of them is generated based on their directory.

This bug was fixed with the help of Mohammad Akhlaghi.

This fixes bug #62563.