developer-build script: Check the existance of the temporary directory
Until now, the temporary directory was set depending on the operating
system (macOS or GNU/Linux). However, it may happen that an operating
system is designed to not have those directories, or an operating system is
used that we hadn't accounted for!

With this commit, the assumption on the operative system has been
removed. Now, the existence of the two possible temporary directories
('/dev/shm' and '/tmp') are checked; and they are only used if they
exist. If none of them exist, then the current directory (.) is considered.

By default, a folder with the name 'gnuastro-HASH' is generated. So, in
order to Git ignore this folder, 'gnuastro-*' has been added to
'.gitignore'.