Lib (type.h): properly reading sci. notation (like 1e5) as integers
Until now, when Arithmetic (or column arithmetic in Table) were given
integers that were written in scientific format (like '1e5'), it would
crash by a bug report: saying that while the number could be read as a
floating point, it can't be read as an integer.

With this commit, we have added a check after the attempt to read the
number as an integer: if the character that couldn't be read as an integer
was 'e', then we simply use the floating point value and write it into the
integer value (instead of re-reading the number).

In the process, I also noticed that until now, we weren't actually checking
if the integer actually falls within the range of 64-bit integers! Without
this check, overflow would occur, and the results were unreasonable! With
this commit, that issue has also been addressed.

This bug was reported by Elham Saremi.

This fixes bug #62564.