Library (type.c): corrected typo in comments
Until now, in the previous commit, I had simply used `~` before the numbers
to signify that the scientific notation is approximate. In the process, I
had forgot to put a '-' (negative sign) before the minimum 64-bit signed
integer value. Also, in Commit d413546d0, we had forgot to add a notice for
fixing bug #62563 in the 'NEWS' file!

With this commit, to avoid confusing '~' with '-', 'almost' is used
instead, and the minimum value now correctly has a '-'. Also, the NEWS file
has been updated with a notice of the bug fix.