psf-stamp: assuming a centered position by default
Until this commit, when the user didn't specify the coordinates to generate
the stamp, the script complained about not having the coordinates of the
desired position. However, this forced to the user to provide the position
of the object even when it was already centered on the image.

With this commit, the script has changed and now it assumes that the object
of interest is in the center of the image (if no coordinates are provided
with --center). As a consequence, when the object is aready centered on the
image, it is not necessary to explicitly provide the coordinates.