psf-stamp: keeping meta-data into the output image header
Until this commit, some useful meta-data used for constructing the stamp
were not saved. But in some scenarios it is important to have that data
because it is necessary in later steps.  More in particular, for this
script it is useful to have the clump and objects labels that were used to
not mask the object. Also the normalization value computed to divide the
whole image.

With this commit, that metadata is saved into the header of the output
image. As a consequence, the user is now able to read that data from the
header using the keywords 'OLABEL', 'CLABEL', and 'NORMVAL'.