psf-stamp: same output stamp size than the input image by default
Until this commit, the script complained about not having the size of the
output stamp if the user didn't provide it. However, this was anoying when
the user just wanted to obtain a stamp with the same size as the original
input image.

With this commit, several changes have been added in order to obtain the
output stamp with the same size than the input image by default. Now it is
not necessary to explicitly provide this information. Moreover, it is also
possible to specify different dimensions for X and Y in order to have
output stamps that are not squares.