Logo: first public draft of Gnuastro's logo
Until now, Gnuastro didn't have any logo. There was only a very crudely
created image that placed the words "GNU Astronomy Utilities" beside the
GNU logo.

With this commit, that old file has been deleted and it is replaced by a
newly designed and original abstract logo. A description of that logo is
also added in the book, and it is placed on the first (title) page of the
PDF book to be immediately visible.

The concept behind this logo was created after several iterations with
Mohammad Akhlaghi.