Configuration: checking if static libraries can be built
Until this commit, Gnuastro didn't check if the static libraries could be
built. However, it may happen that in some systems it is not possible and
then a crash occurs. This is actually what happens in macOS.

With this commit, this problem has been fixed by checking if the compiler
is able to build the static libraries. Now it checks if static libraries
can be built. If it is possible, then it will do it. If it is not possible,
then no static libraries will be created and a warning is printed as an
output.

This fix has been done with the help of Mohammad Akhlaghi, and it fixes the
bug #62590 (https://savannah.gnu.org/bugs/index.php?62590).