Arithmetic: single number written to file when --output is called
Until now, when Arithmetic's final output was a single number and the user
had given an output file name with '--output', Arithmetic would ignore the
request and not make any output file: it would continue printing the value
on the standard output. To write the output in a file, the only solution
was to use the '--onedasimage' option. But in that case, there was no way
to write it as a table (as with all other 1D outputs).

With this commit, a new internal variable has been added to Arithmetic that
will let it know if the user actually called '--output' or not. With this
variable, Arithmetic can behave as expected: when '--output' isn't given,
it will print a single-element output on standard output. Otherwise, it
will respect the user's request and write the output in a file.

This bug was found by Raul Infante-Sainz.

This fixes bug #62597.