Crop: new option to print single-pixel's output to stdout
Until now, Crop would always only create FITS images. However, when the
output only has a single pixel, it is highly likely that the user just
wanted to read that pixel's value. But doing this from a FITS file is
annoying (requires creation of a file for every coordinate/pixel they
want).

With this commit, Crop now has a new option called '--oneelemstdout' (or in
short '-t'). When the output has a single pixel and this option is called,
then crop will print the value on the standard output instead of a FITS
file.

This feature was suggested by Raul Infante-Sainz.