PSF scripts: removing the --corewidth option to simplify the scripts
Until now, the PSF scripts were obtaining the clump and object labels
internally by cropping a small region of the segment image. These labels
are the one that should not be masked. The user could specify the size of
that cropped image with the parameter '--cropwidth'.  However, this
introduced several steps and complexities in the scripts. For example, if
the region was larger than one pixel, the unique value was needed to be
computed.

With the recent completion of the task #16218, Crop has a parameter that
provides the pixel value of a cropped image when it is of one pixel size.
This is very convenient for obtaining the clump and object labels at a
given position, which is exactly the task that the scripts need.

With this commit, the scripts use this new feature to crop 1-pixel image
and obtain its value directly using the new Crop's option
('--oneelemstdout'). As a consequence, the previous function that was
designed for computing this values is not needed any more. In the same way,
the '--corewidth' option has been removed. Now the cropped image to obtain
the object and clumps labels from the segment image is of 1-pixel size.