PSF scripts: homogenizing the way errors and warnings are printed
Until now, the different PSF scripts have different ways of printing
warning and error messages on the standard output (command-line). Some of
them use the "cat<<EOF [text] EOF" and other scripts use simply "echo
text". However, it is better to print all messages following the same
convention.

With this commit, all messages are printed following the same standard:
"cat<<EOF [text] EOF". Moreover, a label indicating if the message is a
warning (WARNING:) or an error (ERROR:) is also printed at the start of the
message. Errors cause the stop of the execution, while warnings are just
useful information. Warnings are not printed if the user uses the option
'--quiet'.