Query: Added short version of Gaia DR3 in VizieR
Until now, we only had a short name for the main Gaia database of ESA. But
VizieR also mirrors Gaia data and in some cases, its easier (with more
standard column names for example!).

With this commit the short name 'gaiadr3' has beed added as a VizieR
dataset, so users don't have to remember (or query!) the low-level VizieR
name.