psf-select-stars: more clear summarizing message
Until now, the progress message which had been printed in the command line
at the end of the 'psf-select-stars' was not clear and confusing: it said
"Number of stars downloaded from gaia"; but the reported number was much
less than number of stars input catlaog (which could have been downloaded
from Gaia). The reported nuber was actually the number of stars in wider
magnitude range!

With this commit, the progress message has been changed to "Number of stars
in wider mag-range", which is clear and describes that number of stars are
in wider magnitude.