Table: --head and --tail corrections moved to where they are applied
Until now, the sanity checks of '--head' and '--tail' were done at the very
start of Table, before any processing was done on it! As part of those
sanity checks, Table would make sure that the values given to these two
options were not larger than the input table's size. It would then adjust
their values if they were larger.

This caused a problem when later steps would limit the number of rows! For
example through the '--range' option. Causing the user to see that the
'--range' option has been ignored for example!

With this commit, the necessary corrections to '--head' and '--tail' are
now done in the same step that they are applied. This fixes the problem
above.

In the process, I also noticed that the formatting sanity checks for the
'--rowlimit' option are done in the 'ui_preparations' function, while the
tests don't depend on the values of the input table! So they are now placed
in the 'ui_read_check_only_options' function which is the proper place for
these tests.

This bug was reported by Sepideh Eskandarlou.

This fixes bug #62636.