Library (arithmetic.h): stitch and noblank are two new operators
Until now, there was no easy way to stitch multiple datasets together in
Gnuastro (for example when adding the separate CCD amplifiers into one
image), or a way remove blank elements of a dataset.

With this commit the two new 'stitch' and 'noblank' operators have been
added to Gnuastro's arithmetic library (and thus the Arithmetic program)
for the two operations above.

In the process, I noticed that the 'unique' operator wasn't defined in the
library, but only in the Arithmetic program! With this commit, it has been
moved into the library to be generally available.