Book: minor typos fixed in newly added section on stitching
Until now, there were a few small typos in the section describing the
'stitch' operator of Arithmetic.

With this commit, they are corrected. In the process, I also noticed that
the descriptions of the Arithmetic library in the book's source are still
not in the new format of one sentence per line! Those parts have also been
corrected to fit the common format.