Library (arithmetic.h): new operators for Janskys to counts and mag
Until now, there was only a single operator for converting counts to
Janskys, but there weren't any operators to convert Janskys to
counts. Also, there were no functions to convert Janskys to AB magnitude.

With this commit the following three operators have been added for the
missing operations mentioned above: 'jy-to-counts', 'mag-to-jy' and
'jy-to-mag'.