Table: adding new rows not changing the column's gal_data_t pointer
Until now, when adding new rows from another file, we were simply making a
whole new table, copying the contents into that table and freeing the
'gal_data_t' pointers of the original columns in the end.

However, this caused a crash when the column pointers were being used in
later steps (for example column arithmetic, which does its preparations in
'ui.c', which include preserving the pointers to the columns): after adding
new rows, the original 'gal_data_t' of the column was freed!

With this commit, the 'gal_data_t' of the columns don't change after adding
rows! We still do allocate a temporary 'gal_data_t', but instead of
replacing the new 'gal_data_t', we simply free the contents, and replace
with the new contents (thus preserving the 'gal_data_t').

This fixes bug #62662.