Configuration: incremented the library SONAME version to 16
Until now, the library SONAME version was 15 (which was used for Gnuastro
version 0.17). However, for for Gnuastro 0.18, this should be incremented
by one because we have made changes in the library. This was usually done
shortly after a new release, but I forgot to do it after version 0.17!

With this commit, it has been incremented.