Table: column arithmetic can now use newly added columns
Until now, column arithmetic could only use columns of the main input, not
any of the columns that were added with '--catcolumnfile'! This was
annoying when you wanted to do some operations based on mixing two sets of
columns.

With this commit, arithmetic can now work on the final set of columns. To
do this, it was necessary to make some modifications to the core structure
for keeping arithmetic information.

In the process, I also corrected the way that a suffix is added to columns
after '--catcolumnfile': until now a suffix would always be added! But now,
a suffix is only added if the same name already exists in the table.

This fixes bug #62674.