Library (arithmetic.h): loading external columns and random from hist
Until now, Gnuastro would only use the Gaussian, Poisson and Uniform
distributions of the GNU Scientific Library (GSL) for random number
generation. However, in some scenarios it is necessary to extract random
values from a custom distribution and there was no way to do it easily.

One of the hinderances was this: it was not possible to easily import
external columns into the arithmetic commands of the Arithmetic or Table
programs. In Arithmetic, all input files should be images and in Table,
they columns have to be in one of the given files, and have to have the
same number of rows. However, inserting custom column was necessary to load
custom-length columns from a table that are only used by the operator and
don't depend on the size of the main input to be operated on (to define the
custom distribution in the example above).

With this commit, a new 'load-col-' operator has been defined in the
Arithmetic library (that is available to both the Arithmetic and Table
programs). This operator can be used to load any column into the processing
stack, and is used by the random histogram distribution to load the
histogram.

Having that operator in place, it was possible to define two new arithmetic
operators in Gnuastro: 'random-from-hist-raw' and 'random-from-hist'. The
first one simply returns a random selection of bin centers that have a
distribution similar to the given histogram. The second uses a uniform
distribution within each bin to random select values within each bin (so
they aren't descrete any more, but still flat if the bins are wide).

Adding the 'random-from-hist' operator was suggested by Raul Infante-Sainz.