psf-stamp: ignore segmentation when star is out of image
Until now, when the stars were outside the field of view, the label of
clump and object were empty and the final stamp would not be
produced. Because in the mask step 'astarithmetic' was looking for the
label of clumps and objects, while these label were empty.

With this commit, when the clump and object are empty the segmentation
image is ingnored.

Also, after pulling the most recent commit, I needed to clean the
'doc/gnuastro-figures' directory and noticed that the 'clean' rule for
'doc/plotsrc/Makefile' was deleting the Gnuastro logo and the README file
of that directory. The 'clean' rule was therefore updated to keep certain
files before deleting others.

This commit was done with the help of Mohammad Akhlaghi.