psf-stamp and psf-scale-factor: corrected comment above changed code
Until now, the comments for the part where we read the clump and object
labels in these two scripts mentioned that the labels are written to a
file. However, since the last commit, that is no longer the case and labels
are just read into memory.

With this commit, those two comments have been removed to avoid
confusion. However, having the clump and object labels can be useful in
debugging, so with this commit, if the script isn't in '--quiet' mode, it
will print the clump and object labels it found to the standard output.

This was pointed out by Raul Infante-Sainz.