psf-scale-factor and psf-stamp: script name in debugging message
Until now (previous commit), the message that would print the clump and
object labels didn't include the script name. It is important that every
program printing messages (like warnings or errors), identify itself in at
the start of the messages.

With this commit, a '$scriptname: ' has been added at the start of the two
'echo' outputs so the user knows which program is printing it.