radial-profile: --stdhdu to MakeCatalog only if given STD image
Until now, the standard deviation (STD) hdu was 1 by default ('--stdhdu=1'
was directly passed to MakeCatalog). So, if no '--instd' was given to the
radial profile script (and therefore no STD crop image was created),
MakeCatalog would use first hdu of the values image as the standard
deviation image!

As a result of this, when a requested measurement column needed the
standard deviation image and the user had forgot to use '--instd', the
radial profile script would set MakeCatalog to use the values file as the
standard deviation!

With this commit, the '--stdhdu=1' option is only passed to MakeCatalog
when an '--instd' file has been given to the radial profile script. Note
that we do crop the standard deviation image before calling MakeCatalog, so
like the values image, its HDU is always one.

This fix was done with the help of Mohammad Akhlaghi.

This fixes bug #62678.