MakeProfiles: --kernel normalized for any zeropoint
Until now, the output of the '--kernel' option was only normalized (the sum
of its pixels were zero) for a zeropoint of 0.0. When a different zeropoint
was used, the output kernel would not be normalized.

With this commit, the final magnitude of the output kernel is set to be
equal to the zeropoint (unless the '--mcolisbrightness' option is given; in
this case, it is set to 1.0).

In the process, I also noticed that with '--magatpeak' the output of
'--kernel' will not be normalized (because the peak value is set to
1.0!). Therefore, if '--magatpeak' is given with '--kernel', MakeProfiles
will abort and print a message to tell the user that these two options
can't be called together.

This bug was found with the help of Raul Infante-Sainz.

This fixes bug #62679.