MakeProfiles: WCS-options ignored even if background doesn't have WCS
Until now, when an image was given to the '--background' option of
MakeProfiles and it didn't have WCS, MakeProfiles would fall back to using
the WCS-options to define a WCS for its output! This is not the expected
behavior and could cause many confusions (as mentioned below).

With this commit, in such cases, the WCS-options are not used and a warning
is printed to let the user know about it.

This bug was found with the help of Raul Infante-Sainz, while investigating
why the radial profile script produces surface brightness measurements even
though the input didn't have any WCS! After investigating the steps, the
culprit was found to be this bug in MakeProfiles.

This fixes bug #62680.