Book: highlighted that MakeNoise's background is per-pixel
Until now, there wasn't enough emphasis in various parts of the book that
in the current implementation, the background value expected by the Poisson
noise generated by MakeNoise or Arithmetic's 'mknoise-poisson' operator is
_per-pixel_, not per arcsec^2 (as is usually reported in observatories for
example).

With this commit, the relevant parts of the manual have been edited/updated
to highlight this important point and hopefully avoid confusion.

Task #16222 (see link [1] below) has been defined to hopefully add the
necessary converter operands for easy conversion in the future.

[1] https://savannah.gnu.org/task/index.php?16222