Arithmetic: name, units and comments of popped operands freed
Until now, in the Arithmetic program, the possibly existing name, units and
comments of a dataset were only freed upon reading a new dataset from a
file. Not when popping an existing dataset from the stack. As a result,
when running a command like below, Arithmetic would crash:

   astarithmetic 50 10 2 makenew set-n n --output new.fits

The reason for this was that the dataset coming from 'makenew' isn't read
from a file and the 'makenew' operator would set a name, units and comments
for its output! On the other hand, the 'set-' operator uses the name of a
dataset to identify which dataset should be used. This conflict caused a
crash!

With this commit, the steps to remove the name, units and comments from a
dataset are now done generally (both when reading from a file, and when
popping from the stack. On the other hand, the extra metadata inserted by
'makenew' has been removed (Arithmetic is about manipulating data, not
metdata!).

This bug was reported by Raul Infante-Sainz.

This fixes bug #62683.